requirements.txt
mysqlclient==2.1.1
python-dotenv==0.21.0
